Intel 360/requirements.txt
streamlit 
transformers
json 
spacy
google.generativeai 
PyPDF2
torch